ref: deprecated useless description methods
